fix: restore UI coverage after submit response contract
Mergeability: exclude generated OpenAPI types from Vitest coverage,
split submitWork tests to satisfy describe line limits, and extend API
tests for extended 201 response fields and session-scoped routes.
