+++ ui/src/api/work/api.test.ts
@@ -1,48 +1,8 @@
-import {
-  SubmitWorkAPIError,
-  submitWork,
-} from "./api";
+import { submitWork } from "./api";
 
-describe("submitWork", () => {
+describe("submitWork routing", () => {
   afterEach(() => {
     vi.unstubAllGlobals();
-  });
-
-  it("returns extended submit work identifiers from a 201 response", async () => {
-    const fetchMock = vi.fn().mockResolvedValue(
-      new Response(
-        JSON.stringify({
-          accepted: true,
-          name: "Driver review",
-          requestId: "request-1",
-          traceId: "trace-story",
-          workId: "work-driver-review",
-          workTypeName: "story",
-        }),
-        {
-          headers: {
-            "Content-Type": "application/json",
-          },
-          status: 201,
-        },
-      ),
-    );
-    vi.stubGlobal("fetch", fetchMock);
-
-    await expect(
-      submitWork({
-        name: "Driver review",
-        payload: "Review the runtime failure.",
-        workTypeName: "story",
-      }),
-    ).resolves.toEqual({
-      accepted: true,
-      name: "Driver review",
-      requestId: "request-1",
-      traceId: "trace-story",
-      workId: "work-driver-review",
-      workTypeName: "story",
-    });
   });
 
   it("posts to the explicit default-session work route and returns the accepted trace id", async () => {
@@ -76,35 +36,6 @@ describe("submitWork", () => {
     );
   });
 
-  it("falls back to a generic message when the server returns an unstructured error", async () => {
-    vi.stubGlobal(
-      "fetch",
-      vi.fn().mockResolvedValue(
-        new Response("server exploded", {
-          headers: {
-            "Content-Type": "text/plain",
-          },
-          status: 500,
-          statusText: "Internal Server Error",
-        }),
-      ),
-    );
-
-    await expect(
-      submitWork({
-        name: "Driver review",
-        payload: "Review the runtime failure.",
-        workTypeName: "story",
-      }),
-    ).rejects.toEqual(
-      new SubmitWorkAPIError({
-        message: "Dashboard submission failed. Try again in a moment.",
-        status: 500,
-        statusText: "Internal Server Error",
-      }),
-    );
-  });
-
   it("posts an explicit empty payload through the default-session scoped route without dropping the submit-work contract field", async () => {
     const fetchMock = vi.fn().mockResolvedValue(
       new Response(JSON.stringify({ traceId: "trace-story" }), {
@@ -131,35 +62,6 @@ describe("submitWork", () => {
           payload: "",
           workTypeName: "story",
         }),
-        method: "POST",
-      }),
-    );
-  });
-
-  it("posts to the session-scoped work route when a non-default session is selected", async () => {
-    const fetchMock = vi.fn().mockResolvedValue(
-      new Response(JSON.stringify({ traceId: "trace-story" }), {
-        headers: {
-          "Content-Type": "application/json",
-        },
-        status: 201,
-      }),
-    );
-    vi.stubGlobal("fetch", fetchMock);
-
-    await expect(
-      submitWork(
-        {
-          name: "Session review",
-          payload: "Review the beta session queue.",
-          workTypeName: "story",
-        },
-        { sessionID: "session-beta" },
-      ),
-    ).resolves.toEqual({ traceId: "trace-story" });
-    expect(fetchMock).toHaveBeenCalledWith(
-      "/factory-sessions/session-beta/work",
-      expect.objectContaining({
         method: "POST",
       }),
     );
